Update README.md: Add "TSV Files" section
If this pull request is merged, it will add a section to the README giving a brief explanation to how to specify the tab character on the command line. Whilst technically this is not about `csv2md` itself, I think the shell syntax required is non-obvious, so users will benefit from the information. TSV is a fairly common file format.